pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai",
    "langgraph",
    "python-dotenv",
]
